+++ source/extensions/isaacsim.replicator.examples/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.10] - 2025-04-17
+### Changed
+- changed add_update_semantics to add_labels
+
 ## [1.1.9] - 2025-04-09
 ### Changed
 - Update all test args to be consistent