+++ source/extensions/isaacsim.sensors.camera/docs/CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+
+## [1.0.1] - 2025-04-17
+### Changed
+- changed add_update_semantics to add_labels
+
 ## [1.0.0] - 2025-04-09
 ### Changed
 - Updates Camera API to account for new lens distortion schemas